Copy result automatically to the clipboard
Several changes were made on code and on UI elements. The second textbox
were totally removed and instead we copy the result to the clipboard of
the user when the conversion succeds.

The new shortcut to invoke the screen is `Command + Option + S`, not
sure we will maintain it, but as the user can freely change it from the
IDE config, I wouldn't botter that much.

Acked-by: Airton Zanon <airtonzanon@gmail.com>
Acked-by: Alexandre Eher <alexandre.eher@werkspot.nl>
Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>